--- packages/react/src/components/Radio/Group/index.ts
@@ -1,2 +1,0 @@
-export { Group } from './Group';
-export type { GroupProps } from './Group';
