+++ packages/react/src/components/form/Radio/Group/index.ts
@@ -1,0 +1,2 @@
+export { RadioGroup as Group } from './Group';
+export type { RadioGroupProps as GroupProps } from './Group';
